Move IMessageSerializer to KnightBus.Messages
Move all packages to netstandard2.1
