+++ KnightBus.sln
@@ -80,7 +80,7 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution 
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "KnightBus.Shared.Tests.Integration", "knightbus\tests\KnightBus.Shared.Tests.Integration\KnightBus.Shared.Tests.Integration.csproj", "{7E533C8D-75F9-46A9-B1F0-186E6EBE6A84}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "KnightBus.ProtoBufNet", "protobuf-net\src\KnightBus.ProtoBufNet\KnightBus.ProtoBufNet.csproj", "{2A5317AA-87FE-4EDE-86EC-D52FC29998C6}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "KnightBus.ProtoBufNet", "knightbus-protobuf-net\src\KnightBus.ProtoBufNet\KnightBus.ProtoBufNet.csproj", "{E049A95F-8602-43A5-8FC9-06B66CD4F0D0}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -208,10 +208,10 @@ Global
 		{7E533C8D-75F9-46A9-B1F0-186E6EBE6A84}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{7E533C8D-75F9-46A9-B1F0-186E6EBE6A84}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{7E533C8D-75F9-46A9-B1F0-186E6EBE6A84}.Release|Any CPU.Build.0 = Release|Any CPU
-		{2A5317AA-87FE-4EDE-86EC-D52FC29998C6}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{2A5317AA-87FE-4EDE-86EC-D52FC29998C6}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{2A5317AA-87FE-4EDE-86EC-D52FC29998C6}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{2A5317AA-87FE-4EDE-86EC-D52FC29998C6}.Release|Any CPU.Build.0 = Release|Any CPU
+		{E049A95F-8602-43A5-8FC9-06B66CD4F0D0}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{E049A95F-8602-43A5-8FC9-06B66CD4F0D0}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{E049A95F-8602-43A5-8FC9-06B66CD4F0D0}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{E049A95F-8602-43A5-8FC9-06B66CD4F0D0}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
