+++ knightbus-azureservicebus/src/KnightBus.Azure.ServiceBus/ServiceBusConfiguration.cs
@@ -1,4 +1,5 @@
 using KnightBus.Core;
+using KnightBus.Messages;
 
 namespace KnightBus.Azure.ServiceBus
 {
