+++ knightbus-microsoft-dependencyinjection/src/KnightBus.Microsoft.DependencyInjection/KnightBus.Microsoft.DependencyInjection.csproj
@@ -1,7 +1,7 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
-    <TargetFramework>netstandard2.0</TargetFramework>
+    <TargetFramework>netstandard2.1</TargetFramework>
     <GeneratePackageOnBuild>true</GeneratePackageOnBuild>
     <Authors>BookBeat</Authors>
     <Description>Microsoft.Extensions.DependencyInjection IoC for KnightBus</Description>
