+++ knightbus-protobuf-net/src/KnightBus.ProtoBufNet/KnightBus.ProtoBufNet.csproj
@@ -1,7 +1,7 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
-    <TargetFramework>netstandard2.0</TargetFramework>
+    <TargetFramework>netstandard2.1</TargetFramework>
     <GeneratePackageOnBuild>true</GeneratePackageOnBuild>
     <Authors>BookBeat</Authors>
     <Description>Protobuf serialization for KnightBus</Description>
@@ -15,7 +15,9 @@
     <PackageTags>knightbus;grpc;protobuf</PackageTags>
     <LangVersion>latest</LangVersion>
   </PropertyGroup>
-
+  <ItemGroup>
+    <None Include="knighbus-64.png" Pack="true" Visible="false" PackagePath="" />
+  </ItemGroup>
   <ItemGroup>
     <PackageReference Include="protobuf-net" Version="3.0.101" />
   </ItemGroup>
