+++ knightbus-redis/src/KnightBus.Redis/LostMessageBackgroundService.cs
@@ -4,6 +4,7 @@ using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
 using KnightBus.Core;
+using KnightBus.Messages;
 using KnightBus.Redis.Messages;
 using StackExchange.Redis;
