+++ knightbus-redis/src/KnightBus.Redis/RedisConfiguration.cs
@@ -1,4 +1,5 @@
 ﻿using KnightBus.Core;
+using KnightBus.Messages;
 
 namespace KnightBus.Redis
 {
