+++ knightbus-redis/src/KnightBus.Redis/RedisQueueClient.cs
@@ -5,6 +5,7 @@ using System.Runtime.CompilerServices;
 using System.Threading;
 using System.Threading.Tasks;
 using KnightBus.Core;
+using KnightBus.Messages;
 using KnightBus.Redis.Messages;
 using StackExchange.Redis;
