+++ knightbus/tests/KnightBus.Shared.Tests.Integration/KnightBus.Shared.Tests.Integration.csproj
@@ -1,7 +1,7 @@
 <Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
-    <TargetFramework>netstandard2.0</TargetFramework>
+    <TargetFramework>netstandard2.1</TargetFramework>
     <Authors>BookBeat</Authors>
     <Copyright>Copyright © BookBeat 2020</Copyright>
     <PackageLicenseExpression>MIT</PackageLicenseExpression>
